+++ R/get-network.R
@@ -74,8 +74,14 @@
 #' @seealso [oe_get()]
 #'
 #' @examples
+#' \dontshow{
+#'   its = file.copy(
+#'     from = system.file("its-example.osm.pbf", package = "osmextract"),
+#'     to = file.path(tempdir(), "test_its-example.osm.pbf"),
+#'     overwrite = TRUE
+#' )}
 #' # default value returned by OSM
-#' its = oe_get("ITS Leeds", quiet = TRUE)
+#' its = oe_get("ITS Leeds", quiet = TRUE, download_directory = tempdir())
 #' plot(its["highway"], lwd = 2, key.pos = 4, key.width = lcm(2.75))
 #' # walking mode of transport
 #' its_walking = oe_get_network("ITS Leeds", mode = "walking", quiet = TRUE)
@@ -83,6 +89,10 @@
 #' # driving mode of transport
 #' its_driving = oe_get_network("ITS Leeds", mode = "driving", quiet = TRUE)
 #' plot(its_driving["highway"], lwd = 2, key.pos = 4, key.width = lcm(2.75))
+#'
+#' # Remove .pbf and .gpkg files in tempdir
+#' # (since they may interact with other examples)
+#' file.remove(list.files(path = tempdir(), pattern = "(pbf|gpkg)", full.names = TRUE))
 oe_get_network = function(
   place,
   mode = c("cycling", "driving", "walking"),
@@ -93,8 +103,8 @@ oe_get_network = function(
   oe_get_options = switch(
     mode,
     cycling = load_options_cycling(place),
-    walking = load_options_walking(place),
-    driving = load_options_driving(place)
+    driving = load_options_driving(place),
+    walking = load_options_walking(place)
   )
 
   # Check the other arguments supplied by the user
@@ -209,7 +219,8 @@ load_options_driving = function(place) {
     highway NOT IN (
     'abandonded', 'bus_guideway', 'byway', 'construction', 'corridor', 'elevator',
     'fixme', 'escalator', 'gallop', 'historic', 'no', 'planned', 'platform',
-    'proposed', 'cycleway', 'pedestrian', 'bridleway', 'path', 'footway'
+    'proposed', 'cycleway', 'pedestrian', 'bridleway', 'path', 'footway',
+    'steps'
     )
     AND
     (access IS NULL OR access NOT IN ('private', 'no'))